Refactor execution.Scheduler tests to prevent import loops
