+++ execution/scheduler_ext_exec_test.go
@@ -1,4 +1,4 @@
-package execution
+package execution_test
 
 import (
 	"encoding/json"
